requirements.txt
asgiref==3.7.2
Django==5.0.3
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-froala-editor==4.1.4
django-js-asset==2.2.0
django-stdimage==6.0.2
pillow==10.2.0
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
